Change library name
android.test.base will be changed to java sdk library, and this library
can not be used as a static library. So this module uses
android.test.base_static instead of android.test.base to behave the same
as the existing one.

Bug: 77577799
Test: make -j

Merged-In: Iddc1e537deab202ad8bcaacbef7613191773e6f3
Change-Id: Iddc1e537deab202ad8bcaacbef7613191773e6f3